Add transaction extension function to SQLiteDatabase
The transaction boiler-plate code that is necessary for each transaction
with the Android SQLiteDatabase class has been modularized into an
extension function to the class. This extension function takes an argument
that is a extension function lambda and executes that function with the
necessary transaction handling (in this case the extension function
lambda is insert() but it could easily be delete() or any other DB
operation).

The function is made more generic by handling generic type T allowing
for the lambda passed to transaction() to return its own type of return
value. In the example this is ID which is a long, but it could
easily be any other return type including Unit (a Kotlin place holder for
no return type).

Finally, the function is declared inline which avoids overhead of making
anonymous classes for the lambda function at execution time. We essentially
gain all the coding benefit of a robust and generic extension function
that aids in code accuracy and readability without taking a performance
penalty.

Signed-off-by: Austin Christ <maximuscato@gmail.com>